Move process module into scheduler module
